#917 refactor: remove shared search barrel
