+++ src/components/landing-page/hero-section/HeroSection.tsx
@@ -1,4 +1,4 @@
-import { SearchBar } from "@/features/search/common";
+import { SearchBar } from "@/features/search/common/components/SearchBar.tsx";
 import { H1 } from "@/components/typography/H1.tsx";
 import { Card } from "@/components/ui/card.tsx";
 import { Trans, useTranslation } from "react-i18next";
