+++ src/components/search/SearchFilters.tsx
@@ -7,8 +7,9 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { MerchantFilters } from "@/components/search/filters/MerchantFilters.tsx";
 import { SellerFilters } from "@/components/search/filters/SellerFilters.tsx";
-import { ShopTypeFilter, useSearchQueryContext } from "@/features/search/common";
+import { ShopTypeFilter } from "@/features/search/common/components/filters/ShopTypeFilter.tsx";
 import { useNavigate } from "@tanstack/react-router";
+import { useSearchQueryContext } from "@/features/search/common/hooks/useSearchQueryContext.tsx";
 import type { SearchFilterArguments } from "@/data/internal/search/SearchFilterArguments.ts";
 import { useCallback, useEffect, useMemo } from "react";
 import { UpdateDateSpanFilter } from "@/components/search/filters/UpdateDateSpanFilter.tsx";
