--- src/features/search/common/index.ts
@@ -1,6 +1,0 @@
-export { ScrollToTopButton } from "./components/ScrollToTopButton.tsx";
-export { SearchBar } from "./components/SearchBar.tsx";
-export { SearchFilterFormProvider } from "./components/SearchFilterFormProvider.tsx";
-export { ShopTypeFilter } from "./components/filters/ShopTypeFilter.tsx";
-export { SearchQueryProvider, useSearchQueryContext } from "./hooks/useSearchQueryContext.tsx";
-export type { SearchFormSchema, SearchType } from "./components/SearchBar.tsx";
