+++ src/integrations/tanstack-query/root-provider.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import type React from "react";
 import i18n from "@/i18n/i18n.ts";
 import { I18nextProvider } from "react-i18next";
-import { SearchQueryProvider } from "@/features/search/common";
+import { SearchQueryProvider } from "@/features/search/common/hooks/useSearchQueryContext.tsx";
 
 export function getContext() {
     const queryClient = new QueryClient({
